+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## XX.XX.XX - 20XX-XX-XX
 
+### PaymentSheet
+* [CHANGED][11112](https://github.com/stripe/stripe-android/pull/11112) Restricted negative initialization amounts.
+
 ## 21.20.2 - 2025-07-09
 
 ### PaymentSheet